tools: spec dep handling
After eveluate the below two possible way:

1. leverage the repodata/*primary.xml.gz for packages 'provides' and 'dependencies'
2. get all necessary info from spec files

Then considered that #1 will actually depend to 'createrepo' and a complete binary
of a release .. the later aka #2 will be self-sustained .. and could also contribute
to modulate spec files to be more clean and robust.

So this commit is continue working on logics for handling meta information from
spec files.

Signed-off-by: Liang Li (Euler) <liliang889@huawei.com>