src/risc-v-kernel: fix the _service file for kernel
The risc-v-kernel pkg is the 'kernel' pkg for oe4rv release. Due to
openSBI / uboot stuff, the kernel pkg itself is not quite as frank
as openEuler-native kernel.

In order to compatiable with OBS and offline build, the _service file
and the source code of risc-v-kernel repo was adjusted accordingly.

Signed-off-by: Liang Li (Euler) <liliang889@huawei.com>